# Conflicts:
#	src/Droid/PrankChat.Mobile.Droid.csproj
#	src/iOS/PrankChat.Mobile.iOS.csproj

+++ src/Core/Models/Data/CreateOrderDataModel.cs
@@ -8,7 +8,7 @@ namespace PrankChat.Mobile.Core.Models.Data
 
         public string Description { get; set; }
 
-        public long? Price { get; set; }
+        public double? Price { get; set; }
 
         public int? ActiveFor { get; set; }
